Move folder to correct path and fix storybook path
